feat(ui): add hkgrotesk light font, set default body fonts, move graphs into dashboard
