+++ src/components/Titles.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const SectionTitle = styled.h2`
+    font-family: HKGrotesk;
+    font-size: 14px;
+    color: #fff;
+    margin-bottom: 15px;
+    font-weight: 600;
+`;
+
+export { SectionTitle };
