+++ src/components/dashboard/DashboardContainer.tsx
@@ -1,7 +1,27 @@
 import * as React from "react";
+import styled from "styled-components";
+import { Flex, Box } from "@rebass/grid";
+import { TransactionGraphPixi } from "../graphs/TransactionGraphPixi";
+import { NetworkGraph } from "../graphs/NetworkGraph";
+import { SectionTitle } from "../Titles";
+
+const Row = styled(Flex)``;
 
 export default class DashboardContainer extends React.Component<{}, {}> {
     public render() {
-        return "Dashboard";
+        return (
+            <>
+                <Row>
+                    <Box width={1 / 2}>
+                        <SectionTitle>Transaction Graph</SectionTitle>
+                        <TransactionGraphPixi />
+                    </Box>
+                    <Box width={1 / 2} style={{ marginLeft: "40px" }}>
+                        <SectionTitle>Network Graph</SectionTitle>
+                        <NetworkGraph />
+                    </Box>
+                </Row>
+            </>
+        );
     }
 }
