+++ src/index.scss
@@ -1,6 +1,7 @@
 @font-face {
     font-family: "Montserrat";
     src: url("./assets/fonts/montserrat/montserrat-regular.ttf");
+    font-weight: normal;
 }
 @font-face {
     font-family: "Montserrat";
@@ -9,7 +10,13 @@
 }
 @font-face {
     font-family: "HKGrotesk";
+    src: url("./assets/fonts/hkgrotesk/hkgrotesk-light.otf");
+    font-weight: 600;
+}
+@font-face {
+    font-family: "HKGrotesk";
     src: url("./assets/fonts/hkgrotesk/hkgrotesk-regular.otf");
+    font-weight: normal;
 }
 @font-face {
     font-family: "HKGrotesk";
@@ -17,8 +24,12 @@
     font-weight: bold;
 }
 
-html {
+html,
+body {
     background-color: #060b24;
+    font-family: Montserrat, -apple-system, BlinkMacSystemFont, “Segoe UI”,
+        “Roboto”, “Oxygen”, “Ubuntu”, “Cantarell”, “Fira Sans”, “Droid Sans”,
+        “Helvetica Neue”, sans-serif;
 }
 
 .tag-list {
